fixed typo
typo in line 114. 
line 67 was not being parsed properly because of a <br> in line 66. Not really sure if it was the <br> tag but removing it seemed to work. 
In the webpage the code block of line 69-81 is not being displayed properly but it displays properly in the preview. not sure why.